+++ src/lib.rs
@@ -446,41 +446,12 @@ pub mod prelude {
     pub use crate::read_only::{ReadOnlyOption, ReadState};
 }
 
-/// Build a logger for tests.
-///
-/// Currently, this is a terminal log. It ensures it is only initialized once to prevent clobbering.
-// This is `pub` so that testing and benching functions can use it.
-// `#[cfg(test)]` doesn't work for benching.
-//
-// TODO: Investigate reporting strategies. Tests are run multithreaded in general, and we'd like to be able to sort through and understand test results.
-#[doc(hidden)]
-pub fn testing_logger() -> &'static Logger {
-    use std::sync::Once;
-    static LOGGER_INITIALIZED: Once = Once::new();
-    static mut LOGGER: Option<Logger> = None;
-
-    unsafe {
-        LOGGER_INITIALIZED.call_once(|| {
-            let decorator = slog_term::TermDecorator::new().build();
-            let drain = slog_term::CompactFormat::new(decorator).build().fuse();
-            let drain = slog_envlogger::new(drain).fuse();
-            let drain = slog_async::Async::new(drain)
-                .chan_size(4096)
-                .overflow_strategy(slog_async::OverflowStrategy::Block)
-                .build()
-                .fuse();
-            LOGGER = Some(slog::Logger::root(drain, o!()));
-        });
-        LOGGER.as_ref().unwrap()
-    }
-}
-
 /// The default logger we fall back to when passed `None` in external facing constructors.
 ///
 /// Currently, this is a `log` adaptor behind a `Once` to ensure there is no clobbering.
 #[doc(hidden)]
 fn default_logger() -> &'static Logger {
-    use std::sync::Once;
+    use std::sync::{Mutex, Once};
     static LOGGER_INITIALIZED: Once = Once::new();
     static mut LOGGER: Option<Logger> = None;
 
@@ -488,20 +459,13 @@ fn default_logger() -> &'static Logger {
         LOGGER_INITIALIZED.call_once(|| {
             let drain = slog_stdlog::StdLog.fuse();
             let drain = slog_envlogger::new(drain).fuse();
-            let drain = slog_async::Async::new(drain)
-                .chan_size(4096)
-                .overflow_strategy(slog_async::OverflowStrategy::Block)
-                .build()
-                .fuse();
-            LOGGER = Some(slog::Logger::root(drain, o!()));
+            LOGGER = Some(slog::Logger::root(Mutex::new(drain).fuse(), o!()));
         });
         LOGGER.as_ref().unwrap()
     }
 }
 
 /// The discard logger is used in `Default::default` implementations.
-///
-/// Currently, this is a `log` adaptor.
 #[doc(hidden)]
 fn discard_logger() -> &'static Logger {
     use std::sync::Once;
