+++ tests/integration_cases/test_raft_snap.rs
@@ -26,9 +26,9 @@
 // limitations under the License.
 
 use crate::test_util::*;
+use crate::testing_logger;
 use harness::Network;
 use raft::eraftpb::*;
-use raft::testing_logger;
 
 fn testing_snap() -> Snapshot {
     new_snapshot(11, 11, vec![1, 2])
