+++ tests/tests.rs
@@ -21,7 +21,7 @@ extern crate slog;
 #[macro_use]
 extern crate lazy_static;
 
-pub use raft::testing_logger;
+pub use harness::testing_logger;
 
 /// Get the count of macro's arguments.
 ///
